Handle case where diffx marks up difference between two images as a few changes within a single w:drawing.
For example, where in v2, the image in v1 was deleted, and replaced with another.  

Word might give both of them the same relId, so the only hint that the image has changed is a couple of del: attributes.

The result is that only the new image is shown, and is not marked up as a change.
      
The expected result is that the old image be shown as a deletion, and the new image shown as an insertion.  
      
Our xslt looks for the tell-tale del: attributes within the w:drawing, and if it finds them, it makes 2 images out of the w:drawing, one of which is marked as an ins, and the other as a del.