Check default content types as well as overriden content types when creating a package.
Previously, only the Override tags in the [Content_Types].xml file were checked to determine what type of OPC package should be created.  This doesn't work when the Content_Types XML file has a Default tag in it, specifying that all files ending in the extension attribute should be treated as the given content type.  Compare:

<Override PartName="/word/document.xml" ContentType="application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml" />
with
<Default Extension="xml" ContentType="application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml"/>