Implement support for em, rem, and ex as length units in various places.
This uses 1rem == 1em == 16px or 1em == medium, and 1ex == 0.5em.
This isn't right, but it's better than nothing.

This also is pretty hacky in that it treats a parsed CSS_NUMBER as a rem.
This is because Flying Saucer is based around CSS2, which doesn't include rem,
so it returns them as plain numbers instead.